changed from Integer to Double

+++ src/main/java/search/com/search/model/dto/ItemsDto.java
@@ -18,8 +18,8 @@ public class ItemsDto {
     private String product;
     private String color;
     private String category;
-    private Integer price;
+    private Double price;
     private String manufacturer;
-    private Integer total;
+    private Double total;
 
 }
